requirements.txt
asgiref==3.4.1
atomicwrites==1.4.0
attrs==21.2.0
boto3==1.18.27
botocore==1.21.27
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
fastapi==0.68.0
h11==0.12.0
httptools==0.2.0
idna==3.2
iniconfig==1.1.1
jmespath==0.10.0
packaging==21.0
pluggy==0.13.1
py==1.10.0
pydantic==1.8.2
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.2
python-dotenv==0.19.0
PyYAML==5.4.1
requests==2.26.0
s3transfer==0.5.0
six==1.16.0
starlette==0.14.2
strip.pysondb==1.5.0
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.6
uvicorn==0.15.0
watchgod==0.7
websockets==9.1
